terrain component; render only visible entities

+++ lib/core/Vector2D.cpp
@@ -5,11 +5,11 @@ Vector2D::Vector2D() :x(0),y(0)
 {
 }
 
-Vector2D::Vector2D(float xy) :x(xy),y(xy)
+Vector2D::Vector2D(int xy) :x(xy),y(xy)
 {
 }
 
-Vector2D::Vector2D(float x, float y) :x(x),y(y)
+Vector2D::Vector2D(int x, int y) :x(x),y(y)
 {
 }
 
@@ -75,7 +75,7 @@ Vector2D& Vector2D::operator/=(const Vector2D& v)
 	return *this;
 }
 
-double Vector2D::distanceTo(const Vector2D& v) const
+int Vector2D::distanceTo(const Vector2D& v) const
 {
 	return sqrt(pow(v.x - x, 2) + pow(v.y - y, 2));
 }
