+++ lib/core/Vector2D.h
@@ -4,12 +4,12 @@
 class Vector2D
 {
 public:
-	float x;
-	float y;
+	int x;
+	int y;
 
 	Vector2D();
-    Vector2D(float xy);
-	Vector2D(float x, float y);
+    Vector2D(int xy);
+	Vector2D(int x, int y);
 	friend Vector2D& operator+(Vector2D& v1, const Vector2D& v2);
 	friend Vector2D& operator-(Vector2D& v1, const Vector2D& v2);
     friend Vector2D& operator*(Vector2D& v1, const Vector2D& v2);
@@ -19,5 +19,5 @@ public:
 	Vector2D& operator*=(const Vector2D& v);
 	Vector2D& operator/=(const Vector2D& v);
 	friend std::ostream& operator<<(std::ostream& stream, const Vector2D& v);
-	double distanceTo(const Vector2D& v) const;
+	int distanceTo(const Vector2D& v) const;
 };
